Move sortResults, remove unused sortGroupedResults
