+++ packages/scrabble-solver/src/state/results/lib.ts
@@ -1,7 +1,7 @@
-import { Result } from '@scrabble-solver/types';
+import { Result, ShowCoordinates } from '@scrabble-solver/types';
 
-import { createRegExp } from 'lib';
-import { CellFilter, GroupedResults } from 'types';
+import { createKeyComparator, createRegExp, createStringComparator, getCoordinates, reverseComparator } from 'lib';
+import { CellFilter, Comparator, GroupedResults, ResultColumnId, Sort, SortDirection } from 'types';
 
 export const groupResults = (
   results: Result[] | undefined,
@@ -48,4 +48,37 @@ export const resultMatchesCellFilter = (result: Result, cellFilters: CellFilter[
   });
 
   return matchesExcludeFilters && matchesIncludeFilters;
+};
+
+const comparators: Record<ResultColumnId, (locale: string, showCoordinates: ShowCoordinates) => Comparator<Result>> = {
+  [ResultColumnId.BlanksCount]: (locale: string) => createKeyComparator('blanksCount', locale),
+  [ResultColumnId.ConsonantsCount]: (locale: string) => createKeyComparator('consonantsCount', locale),
+  [ResultColumnId.Coordinates]: (locale: string, showCoordinates: ShowCoordinates) => (a, b) => {
+    const stringComparator = createStringComparator(locale);
+    const aValue = getCoordinates(a, showCoordinates);
+    const bValue = getCoordinates(b, showCoordinates);
+    return stringComparator(aValue, bValue);
+  },
+  [ResultColumnId.Points]: (locale: string) => createKeyComparator('points', locale),
+  [ResultColumnId.TilesCount]: (locale: string) => createKeyComparator('tilesCount', locale),
+  [ResultColumnId.VowelsCount]: (locale: string) => createKeyComparator('vowelsCount', locale),
+  [ResultColumnId.Word]: (locale: string) => createKeyComparator('word', locale),
+  [ResultColumnId.WordsCount]: (locale: string) => createKeyComparator('wordsCount', locale),
+};
+
+export const sortResults = (
+  results: Result[] | undefined,
+  sort: Sort,
+  locale: string,
+  showCoordinates: ShowCoordinates,
+): Result[] | undefined => {
+  if (typeof results === 'undefined') {
+    return undefined;
+  }
+
+  const createComparator = comparators[sort.column];
+  const comparator = createComparator(locale, showCoordinates);
+  const finalComparator = sort.direction === SortDirection.Descending ? reverseComparator(comparator) : comparator;
+  const sortedResults = [...results].sort(finalComparator);
+  return sortedResults;
 };
